Fix: Error when loading sampled images (#339)
**Description:**

- Load images as whole bitmap and not as buffered source

**Explaination**
Loading AES/GCM encrypted files buffered/sampled. Causes the stream to
skip bytes which is not allowed in AES/GCM.

Since this raises the issue of crashing with very big images again. This
is once of the reasons I want to overhaul the encryption some day to not
use GCM anymore #204

Fixes #330 

- [x] Confirmed working by @prosch88