Fix: Only use photo picker when not using "Delete imported files" (#418)
**Description:**

The only possibility to delete the imported files is not to use the
photo picker.

The photo picker does not expose the real media uri that is needed for
deletion but only a syntetic uri with read only access.

Now the photo picker is used for when read only is enough. And classic
open document is used when the app needs delete access.

Altough:

Now its still possible to select photos and videos even when not using
the photo picker. So no functionality is lost.