Bug: Fix crash with very big images (#419)
**Description:**

Fixes a crash when displaying very big images.

Closes #403 

Explanation:
Normal downsampling is currently not possible. Because of the nature of
AES/GCM (seems like changing the IV to 12 bytes fixes this, but this is
a much bigger topic)

Applied solution: Read all bytes into memory and provide as a buffered
source from memory. This allows downsampling done by coil, since its
already decrypted in memory.