Feature: Biometric Unlock (#502)
This PR introduces biometric authentication as an alternative unlock
mechanism, allowing users to unlock the app with their fingerprint or
other biometrics.

## How it works

The feature can be enabled in the settings. After confirming your
biometric identity. Its enabled.

When opening Photok, you are automatically prompted to unlock it via
biometrics.

## Technical / Encryption 

This feature works by storing a encrypted version of the users password
derived AES key in the apps internal storage.

This key is encrypted (wrapped) by another AES key that is located in
the Android Keystore. The wrapping key is protected by biometric
authentication and can therefore not be used without authentication.

When disabling the feature, the stored key is deleted.

Closes #45