Fix: Wrong file order when importing backup (#556)
This fixes the order of files when importing a backup. It recreates the
original order before assigning new import timestamps.

In a future rewrite of the restore backup UI it should be possible to
configure if the original timestamp should be kept.

Additionally this PR changes the restoring of V1 backups to also first
process to files and then insert all database items.

Fixes: #380