Fix: Reset biometric unlock when changing password
This fixes a bug where no images can be decrypted if unlocking Photok
with biometrics after changing the password.

This is caused by the stored user key is not replaced when reencrypting
data. This also cannot be done since we would need to fingerprint to
unlock the biometric cipher to update the user key. This is the most
pragmatic and simple solution.